Use an explicit locale.
This isn't just an idle warning; while Java always guarantees that the
US locale will be available, and the US locale is always the right one
to use for HTTP headers (which are guaranteed to be US-ASCII), there's
no guarantee that the current _default_ locale will be anything like
that. It's an easy fix, in any case -- the default locale is always
there, and the closest we have to "Just treat it as ASCII".