Fix no warning in log if temp file was not deleted
The method delete() of interface TempFile should message a failed delete
by throwing an Exception. For the default implementation DefaultTempFile
no Exceptions were thrown because the used method delete() of File does
not throw an Exception on failure but returns false instead. The problem
was fixed by checking the return value and throwing an Exception for it.