Fix webserver serving folder named like index file
If there was a directory named like an index file (e.g. "index.html"),
other index files registered later (e.g. "index.htm") were ignored and
the directory was served instead (or an index file in the directory).
This was fixed by now checking if the index file is not a directory.