Add a RoutePrioritizer interface and some implementations of that interface.  There is a Default implementation that mimics the existing code, an InsertionOrder implementation that uses the order that routes are inserted, and a ProvidedPriority implementation that uses the priority that was provided instead of the one that UriResource calculates for itself.  By default routing works the same, but this allows us to extend how routing works.
If you have the following routes:

/general/:param1
/general/:param1/:param2

and you hit /general/abc/123, you would expect the second route to be executed with param1=abc and param2=123, but what actually happens is that the second route is executed with param1=abc/123.

While this could be fixed by inverting priorities so that things with more parameters will end up first in the list.  This ultimately fails, however, for cases like /user/help vs /user/:param1 where the second would then always execute with param1=help.

This fixes it to go entirely based on insertion order.